Split transports.py into modular structure (#2921)
