+++ docs/python-sdk/fastmcp-client-transports-__init__.mdx
@@ -1,0 +1,8 @@
+---
+title: __init__
+sidebarTitle: __init__
+---
+
+# `fastmcp.client.transports`
+
+*This module is empty or contains only private/internal implementations.*
